Move common classes to 'game-models' and 'game-logic' folders
