+++ common/game-models/src/main/java/enchantedtowers/game_logic/CurvesMatchingMetric.java
@@ -1,4 +1,4 @@
-package enchantedtowers.client.components.spell;
+package enchantedtowers.game_logic;
 
 import org.locationtech.jts.geom.Geometry;
