+++ common/game-models/src/main/java/enchantedtowers/game_logic/HausdorffMetric.java
@@ -1,7 +1,9 @@
-package enchantedtowers.client.components.spell;
+package enchantedtowers.game_logic;
 
 import org.locationtech.jts.algorithm.match.HausdorffSimilarityMeasure;
 import org.locationtech.jts.geom.Geometry;
+
+import enchantedtowers.game_logic.CurvesMatchingMetric;
 
 public class HausdorffMetric implements CurvesMatchingMetric {
     @Override
