+++ common/game-models/src/main/java/enchantedtowers/game_models/Spell.java
@@ -1,4 +1,4 @@
-package enchantedtowers.client.components.spell;
+package enchantedtowers.game_models;
 
 import org.locationtech.jts.geom.Coordinate;
 import org.locationtech.jts.geom.Envelope;
